+++ cx-portal-shared-components/src/components/basic/Stepper/StepperItem.tsx
@@ -28,6 +28,7 @@ interface StepperItemProps {
   headline: string
   activeStep: number
   index: number
+  totalSteps?: number
 }
 
 export const StepperItem = ({
@@ -35,6 +36,7 @@ export const StepperItem = ({
   headline,
   activeStep,
   index,
+  totalSteps = 6,
 }: StepperItemProps) => {
   const [backgroundColor, setBackgroundColor] = useState(
     theme.palette.stepper.stepUpcoming
@@ -53,10 +55,11 @@ export const StepperItem = ({
       setDone(false)
     }
   }, [index, activeStep])
+  const width = 100 / totalSteps
   return (
     <Box
       sx={{
-        width: '186px',
+        width: `${width}%`,
         margin: '0px',
         borderBottom: `2px solid ${backgroundColor}`,
       }}